refactored entire codebase and remove unnecessary code
