+++ urban-dashboard/src/components/infoSidebar/index.js
@@ -6,14 +6,15 @@ import { PopulationCard } from '../cards/population';
 import { DataInsightsCard } from '../dataInsights';
 import { Title } from '../title';
 import { ColorMapChart } from '../cards/colorMapChart';
+import { DATASETS } from '../../constants';
 
 import "./index.css";
 
-function DataSnippet({ dataset }) {
+const DataSnippet = ({ dataset }) => {
     let briefSnippet = "data is available for the contiguous United States. Locations indicated are for representative purposes only, based on the top 30 U.S. cities by population size, according to the 2021 U.S. Census Bureau figures and city “Place” boundaries.";
-    if (dataset === "vulcan") {
+    if (dataset === DATASETS.VULCAN) {
         briefSnippet = `Vulcan ${briefSnippet}`
-    } else if (dataset === "gra2pes") {
+    } else if (dataset === DATASETS.GRA2PES) {
         briefSnippet = `GRA²PES ${briefSnippet}`
     }
 
@@ -27,13 +28,13 @@ function DataSnippet({ dataset }) {
 }
 
 // TitlePopulationAndArea is a wrapper for Title, Data Snippet and Population Cards
-function TitlePopulationAndArea({
+const TitlePopulationAndArea = ({
     urbanRegions,
     selection,
     setSelection,
     handleZoomOut,
     dataset
-}) {
+}) => {
     return (
         <div>
             <Title
@@ -42,22 +43,15 @@ function TitlePopulationAndArea({
                 handleZoomOut={handleZoomOut}
             />
 
-            {/* Data Snippet Card */}
-            {!selection && <DataSnippet dataset={dataset} />}
+            {!selection ? <DataSnippet dataset={dataset} />
+                : <PopulationCard selection={selection} urbanRegions={urbanRegions} />}
 
-            {/* Population and Area Card */}
-            {selection && (
-                <>
-                    <PopulationCard selection={selection} urbanRegions={urbanRegions} />
-                </>
-            )}
         </div>
     )
 }
 
-
 // DataInsights plots charts on the info sidebar
-function DataInsights({ dataset, selection }) {
+const DataInsights = ({ dataset, selection }) => {
     return (
         <>
             {selection && (
