--- urban-dashboard/src/components/insights/helper/left.jsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-export function LeftInsights(props) {
-    return (
-        <div className="insights-left">
-            {props.children}
-        </div>
-    );
-}
-
-export function BottomLeftInsights(props) {
-    return (
-        <div className="insights-bottom-left">
-            {props.children}
-        </div>
-    );
-}
-
-LeftInsights.propTypes = {
-    children: PropTypes.oneOfType([
-        PropTypes.arrayOf(PropTypes.element),
-        PropTypes.element.isRequired
-    ])
-}
