--- urban-dashboard/src/components/insights/helper/right.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-export function RightInsights(props) {
-    return (
-        <div className="insights-right">
-            {props.children}
-        </div>
-    );
-}
-
-RightInsights.propTypes = {
-    children: PropTypes.oneOfType([
-      PropTypes.arrayOf(PropTypes.element), 
-      PropTypes.element.isRequired
-    ])
-}
