--- urban-dashboard/src/components/insights/index.css
@@ -1,36 +1,0 @@
-/* parent div of insights should be relative,
- else insight would be applied with repect to body */
-
-.insights {
-    display: block;
-    position: absolute;
-    width: 100%;
-    height: 100vh;
-    top: 11%;
-    color: white;
-}
-
-.insights-left {
-    display: block;
-    position: absolute;
-    left: 1%;
-    top: 0;
-    bottom: 0;
-}
-
-.insights-bottom-left {
-    position: absolute;
-    z-index: 1000;
-    bottom: 15%;
-    left: 1%;
-}
-
-.insights-right {
-    display: block;
-    position: absolute;
-    right: 1%;
-}
-
-.insights>div>div {
-    margin-bottom: 4%;
-}
