+++ urban-dashboard/src/constants.js
@@ -1,0 +1,6 @@
+
+// List of datasets available in this dashboard
+export const DATASETS = {
+    VULCAN: "vulcan",
+    GRA2PES: "gra2pes",
+}
